knative/MongoTriggerFn/app/requirements.txt
fastapi
httptools
kafka-python
pip-chill
python-dotenv
uvloop
watchfiles
websockets

